CIV-7531 : Move function from DeadlineExtensionCalculatorService to CoreCaseDataService
