+++ src/main/java/uk/gov/hmcts/reform/civil/service/citizenui/responsedeadline/DeadlineExtensionCalculatorService.java
@@ -3,9 +3,6 @@ package uk.gov.hmcts.reform.civil.service.citizenui.responsedeadline;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import uk.gov.hmcts.reform.civil.bankholidays.WorkingDayIndicator;
-import uk.gov.hmcts.reform.civil.helpers.CaseDetailsConverter;
-import uk.gov.hmcts.reform.civil.model.CaseData;
-import uk.gov.hmcts.reform.civil.service.CoreCaseDataService;
 
 import java.time.LocalDate;
 
@@ -14,8 +11,6 @@ import java.time.LocalDate;
 public class DeadlineExtensionCalculatorService {
 
     private final WorkingDayIndicator workingDayIndicator;
-    private final CoreCaseDataService coreCaseDataService;
-    private final CaseDetailsConverter caseDetailsConverter;
 
     public LocalDate calculateExtendedDeadline(LocalDate dateProposed) {
         return workingDayIndicator.isWorkingDay(dateProposed)
@@ -23,8 +18,4 @@ public class DeadlineExtensionCalculatorService {
             : workingDayIndicator.getNextWorkingDay(dateProposed);
     }
 
-    public LocalDate getAgreedDeadlineResponseDate(Long caseId, String authorization) {
-        CaseData caseData = caseDetailsConverter.toCaseData(coreCaseDataService.getCase(caseId, authorization));
-        return caseData.getRespondentSolicitor1AgreedDeadlineExtension();
-    }
 }
